erika_cloud/requirements.txt
bottle
peewee
paho-mqtt
python-dateutil
python-dotenv